Section 11: Adding JTW Claims

+++ Section11/resources/item.py
@@ -1,6 +1,6 @@
 import sqlite3
 
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt_claims
 from flask_restful import Resource, reqparse
 from models.item import ItemModel
 
@@ -43,13 +43,17 @@ class Item(Resource):
         return item.json(), 201
 
 
-    # @jwt_required()
+    @jwt_required
     def delete(self, name):
+        claims = get_jwt_claims()
+        if not claims['is_admin']:
+            return {'message': 'Admin privilege required.'}, 401
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
+            return {'message': 'Item deleted'}
 
-        return {'message': 'Item deleted'},404
+        return {'message': 'Item not found'},404
 
     # @jwt_required()
     def put(self, name):
